lrng - reset entropy gathering interface in any case
The interrupt gathering interface diverts raw entropy data from the LRNG
to the interface when a caller listens. Irrespective whether the caller
obtained all data or not, the gathering logic is always reset when the
caller is not listening any more.

The only exception is the boot time data which remains in the buffer
until the buffer is completely emptied. Yet, in this case, the entropy
gathering is not diverting entropy data to the test interface any more.

Signed-off-by: Stephan Mueller <smueller@chronox.de>