backports 4.x: add missing fix
The backports for kernels 4.x miss a patch to make the hash operations
working. The patch sets are tested now on CentOS 8 and 7 to work out of
the box without further changes.

Note, the LRNG on 4.14 and lower uses SHA-1 per default as the
lib/sha256.c code is not present. If a SHA-256 implementation is
intended, this file and its header file needs to be backported.

Reported-by: John Masinter <john@masinter.net>
Signed-off-by: Stephan Mueller <smueller@chronox.de>