LRNG: update per-CPU entropy pool more often during boot
During boot, the performance of the interrupt handler may be a bit
slower if on the other hand entropy is delivered faster. The patch
upates the per-CPU entropy pool more often during boot to ensure entropy
is available for harvesting in due time.

This allows the enlargement of the default per-CPU array to reduce the
number of hash operations at run time where new entropy may not be
needed as often assuming that the DRNG is fully seeded.

This patch increases the performance of the interrupt handler.

Signed-off-by: Stephan Mueller <smueller@chronox.de>