LRNG: only use hash update in interrupt handler
To increase the performance of the interrupt handler, the hash init and
hash final operations are moved to the time when the per-CPU entropy
pool contents are collected. This implies that in the interrupt handler
only hash update is invoked which makes the interrupt handler
significantly faster: the LRNG now operates up to 130% faster in the
interupt handler compared to the legacy /dev/random implementation
depending on the collection size. The default of the collection size is
to support the fastest interrupt handling implementation.

The per-CPU entropy pool is now changed to be the hash state. This
implies that during hash switching or NUMA allocation, the hash state
needs to be updated. This also implies that during collection of the
per-CPU entropy pools, the hash implementation active for each NUMA node
at the given time must be used.

Signed-off-by: Stephan Mueller <smueller@chronox.de>