LRNG: increase entropy pool size config values
The configuration values were set based on the old LFSR model. As we
have a hash-based conditioning now, we can freely choose the entropy
pool size. Thus, the configurable pool size is enlarged:

- allowing small memory constrained devices

- allowing the regular size

- allowing large environments

Note, the larger the entropy pool is, the less impact it has on the
interrupt handler performance.

Signed-off-by: Stephan Mueller <smueller@chronox.de>