LRNG: report available pool size
The LRNG allocates a pool for each CPU. Yet, the pools are only filled
with data from interrupt events on CPUs on which an interrupt is
scheduled. It is possible and even likely that on an SMP system some
CPUs are not used to serve interrupts. Thus, the entropy pools on those
CPUs are dormant and not used.

When reading the maximum pool size, the LRNG now reports the pool size
for all per-CPU entropy pools that are in use.

Signed-off-by: Stephan Mueller <smueller@chronox.de>