LRNG: support disabling of continuous compression
The Linux RNG collects entropy data during each interrupt. For
performance reasons, a amount of entropy data defined by the
LRNG entropy collection pool size is concatenated into an array.
When that array is filled up, a hash is calculated to compress
the entropy. That hash is calculated in interrupt context.

In case such hash calculation in interrupt context is deemed
too time-consuming, the continuous compression operation
can be disabled. If disabled, the collection of entropy will
not trigger a hash compression operation in interrupt context.
The compression happens only when the DRNG is reseeded. This
implies that old entropy data collected after the last
DRNG-reseed is overwritten with newer entropy data instead
of retaining its entropy with the compression operation.

The modification contains the following implied changes:

- Forcefully enable the compression if data array holding the
  entropy data is too small to hold at least as much entropy as
  required to satisfy the security strength of the LRNG

- Report the poolsize correctly depending on the entropy rate present
  with or without the continuous compression rate

- Report the enabled continuous compression state via lrng_type file

- The data array is not memset(0), but each injection of a value into
  the array is preseeded with AND-ing a zero into the data array slot to
  ensure the OR operation is mathematially correct. This allows a
  continuous overwriting of the data array without continuous
  compression and yet retain the old entropy data.

- Update the self-test to "poison" (i.e. dirty) the data array before
  filling it to verify the correctness of the zeroization operation.

Signed-off-by: Stephan Mueller <smueller@chronox.de>