LRNG: update seeding threshold when loading DRNG
When the loading of the DRNG is performed while the LRNG is not fully
seeded yet, the seeding threshold must be updated as well. For example,
if SHA-1 is used as default with 160 bit threshold and loading SHA-256
before reaching the fully seeded level, the LRNG will never be able to
reach the fully seeded level without updating the threshold level.

Reported-by: Matthew McClain <mmcclain@uplogix.com>
Signed-off-by: Stephan Mueller <smueller@chronox.de>