LRNG: merge hash processing into one location
To simplify the code and make the data processing more obvious, the hash
update operation for the auxiliary pool is moved to the location where
the message digest is generated.

This code change also increases the performance as only the digest size
bits are inserted into the auxiliary pool. Peviously it was always 64
bytes which may not be used if a smaller hash than SHA-512 is used.

To avoid unnecessary memset() calls to clear the digest buffer, the
function lrng_pcpu_pool_hash still maintains the digest buffer even
though it is not processed there.

Signed-off-by: Stephan Mueller <smueller@chronox.de>