LRNG: add oversampling support
When using conditioning operations, the entropy rate can only
asymptotically approach full entropy. This means that if the
conditioning function is able to output 256 bits of data, 256 bits of
entropy in that data requires a bit more entropy. This is due to the
fact that random maps will tend to have output collisions, which
reduce the creditable output entropy.

The LRNG applies the value of 64 bits of entropy for this oversampling
strategy. This value is justified in Appendix A.4 of the current
SP800-90C draft, and aligns with NIST's in "epsilon" definition in
this document, which is that a string can be considered "full entropy"
if one can bound the min entropy in each bit of output to at least
1-epsilson, where epsilon is required to be <= 2^(-32).

In addition, a separate oversampling is applied when initially seeding
the DRNG. In order to support the claim that the DRNG is fully seeded,
the LRNG applies at least 128 more bits of entropy. If a DRNG has a
security strength of 256 bits, the initial seeding will provide 384 bits
of entropy. Any subsequent seeding will attempt to seed 256 bits of
entropy. The value of 128 bits are derived from the current draft of
SP800-90C.

The oversampling support is a compile time option.

The following changes are visible:

- all entropy source interface functions use an input parameter of
  requested_bits indicating the amount of entropy that is to be provided

- treat all noise sources identically - if a set of noise sources other
  than the interrupt noise source provide suffient entropy, mark the
  DRNG as fully seeded

- request 64 bits more of entropy from the entropy pools and the aux
  pool than requested to satisfy the first requirement above

- enlarge the seed buffer by 128 bits for each entropy source to provide
  room for the initial seeding of a DRNG

- mark a newly loaded DRNG to not fully seeded when a DRNG switch occurs
  to guarantee it is seeded with 128 more bits

This patch supports full compliance with the current SP800-90C draft.

Signed-off-by: Stephan Mueller <smueller@chronox.de>