LRNG: cap the available entropy during hash switch
When the hash is switched, the available entropy rate is capped to the
old hashes' digest size. This supports the fact that if a switch from,
say, a SHA-256 to SHA-512 will not cause the available entropy value to
duplicate. This is appropriate because the existing entropy was
processed with the SHA-256 conditioning function which only has 256 bits
at a maximum.

Note, when changing from SHA-512 to SHA-256, this change does not do
anything. Yet, this scenario is already caught in the functions
harvesting entropy that cap the available entropy to the digest size of
the currently used hash.

Signed-off-by: Stephan Mueller <smueller@chronox.de>