LRNG: Remove seeding invocation during initialization
When the ChaCha20 DRNG is initialized, we do not need to manually seed
the DRNG. The lrng_drngs_init_cc20 is called before the first ever
generate operation irrespective which LRNG generate interface is
used. By triggering the first seed operation in this function, all
available entropy at that time is used to seed the ChaCha20 DRNG. This
includes the entropy from all entropy sources, including fast and slow
sources, such as the CPU entropy source (e.g. RDSEED).

Signed-off-by: Stephan Mueller <smueller@chronox.de>