LRNG: force reseed after system init
The system initialization call of LRNG rand_initialize invokes the
arch_get_random_seed_long_early which on some architectures may perform
an additional initialization step that is requires to get good entropy
data.

The LRNG forces an initial reseed with lrng_drngs_init_cc20() when the
first request for random numbers arrives. After that, the next reseed is
when the initially required amount of interrupts are received.

It may happen, that rand_initialize is called after the first random
number is generated. In this case, the LRNG initially seeded from the
CPU noise sources without using the arch_get_random_seed_long_early
call. To ensure that a new seeding operation is performed after
arch_get_random_seed_long_early is invoked to guarantee that the LRNG is
seeded from a fully initialized CPU entropy source, a reseed is also
triggered.

Signed-off-by: Stephan Mueller <smueller@chronox.de>