LRNG: make entropy rate of IRQ entropy source configurable
The entropy rate implied by the LRNG for the IRQ entropy source is the
following: it requires 256 interrupt events to generate 256 bits of
entropy. In case this value is not considered to be appropriate, it can
now be altered by the kernel command line option of
lrng_pool.irq_entropy.

This configuration can only be applied at boot time. The LRNG enforces
that the value is at least 256. This implies that using this
configuration the entropy rate can only be lowered.

In case the IRQ entropy source should not be credited with entropy, the
value 0xffffffff must be supplied to the aforementioned option. This
implies that still data from the entropy source is obtained, but it does
not add to the LRNG entropy estimation.

Signed-off-by: Stephan Mueller <smueller@chronox.de>