LRNG: speed up early boot entropy gathering
If an external entropy source provides entropy before the LRNG is fully
seeded, it is now used to speed up reaching that threshold. This is
ensured by lowering the reseeding threshold by the amount of entropy
added by the external source.

In addition, the oversampling rate is only applied if the SP800-90C
compliant mode is chosen by the caller. This makes the threshold setting
consistent with the collection of the entropy.

Signed-off-by: Stephan Mueller <smueller@chronox.de>