LRNG: initialize Jitter RNG during device_initcall
With the latest changes, the Jitter RNG may be called very early in the
boot cycle. On some architectures, an appropriate timer is not available
at that time because the clocksource is initialized with fs_initcall.

Thus, the Jitter RNG is initialized now at device_initcall. At that time
it also informs the LRNG that it can now provide entropy. If the LRNG is
not yet fully seeded, this lowers the reseed thresholds by the amount of
entropy delivered by the Jitter RNG.

Signed-off-by: Stephan Mueller <smueller@chronox.de>