LRNG: update seeding threshold when loading DRNG only if min seeded
The update of the seeding threshold is only needed when the final
entropy threshold is set, i.e. when the LRNG is already minimally
seeded. If the threshold is set before, the minimally seeded threshold
is skipped.

Thus, only update the seeding threshold when loading a DRNG instance
and switching to it if the LRNG is minimally seeded.

Signed-off-by: Stephan Mueller <smueller@chronox.de>