LRNG: insert bootloader entropy directly into the aux pool
The boot loader entropy is provided very early in the boot cycle before
much of the infrastructure of the kernel is available. It is also
provided before the work queue mechanism is initialized. Thus, using
add_hwgenerator_randomness is not advisable.

The change therefore inserts the entropy directly into the auxiliary
pool without using any helper function

Reported-by: Alexander Lobakin <alobakin@mailbox.org>
Signed-off-by: Stephan Mueller <smueller@chronox.de>