LRNG: use work queue after initialization only
It is possible that the LRNG is seeded earlier than the work queue
mechanism is available. When this occurs together with the call to
get_random_bytes before the work queue is initialized, the seeding
operation is stuck. Thus, only use the work queue once it is known that
the work queue system is initialized.

Signed-off-by: Stephan Mueller <smueller@chronox.de>