LRNG: invoke invalidate_batched_entropy after initialization
During very early boot, the CPU masks may not yet be set up properly.
Thus, the code to iterate over present CPUs may not yet be called. Such
code is used in invalidate_batched_entropy. Therefore, ensure that this
code is only called once the LRNG is triggered by the kernel that the
subsystem is available.

Signed-off-by: Alexander Lobakin <alobakin@mailbox.org>
Signed-off-by: Stephan Mueller <smueller@chronox.de>