LRNG: always fill in the time stamp into seed buffer
When insufficient entropy is present, the seed buffer is not filled with
any entropy. The caller may use it for seeding nonetheless. Therefore,
at least a variating factor should be present. Thus, move the addition
of the time stamp into the seed buffer before the check of the available
entropy.

Signed-off-by: Stephan Mueller <smueller@chronox.de>