LRNG: signal end of boot cycle in non-NUMA configuration
The boot cycle is marked to be complete with the flag
all_online_numa_node_seeded. In case of the non-NUMA environment that
was not set and thus an additional but unsuccessful reseed attempt is
tried. This wastes CPU cycles. Therefore, set this flag also in a
non-NUMA setup.

This may imply, however, that NUMA nodes now are a bit later seeded
because the interrupt handler code triggering the pings for new entropy
fires only when the collection pool is full.

Signed-off-by: Stephan Mueller <smueller@chronox.de>