LRNG: set LRNG to non-operational for 90C compliance
When SP800-90C compliance is requested, and the cryptographic primitives
supporting the 90C support are loaded, the LRNG is set back to
non-operational. This implies that the blocking interfaces are gated
again until the SP800-90C seed is added.

The change also removes the collection pool lock for the interrupt
entropy source when switching the DRNG. The reason is that the
collection pool is not touched, but only the hash state is changed. That
hash state is locked with a write lock by the DRNG switch logic. Thus,
the hash state is modified in an atomic fashion.

Signed-off-by: Stephan Mueller <smueller@chronox.de>