LRNG: harden entropy source configuration
The entropy source's entropy rate configuration is changed into a
compile-time option to prevent regular administrators to modify the
settings. It is assumed that common administrators may not have a good
knowledge about the entropy estimaation.

The compile-time options now provide the means to set the applied
entropy rates with the kernel configuration system and make it visible
which values are set.

This implies that the kernel configuration options for the entropy rate
are added. To reduce cludder, the entropy source configuration options
are now moved into a sepaarate configuration item.

To allow integrators or testers to adjust the entropy rates during
development, a compile time option is offered to allow configuring them
at boot/runtime. This kernel configuration option is moved to the
testing section indicating that this option is intended for testing.
Yet, the runtime interfaces are considered to be secure to the extent
that they could be used for production kernels.

The test code is updated to check for the enabling of the configuration
option allowing the change of the entropy rate at boot/runtime.

Signed-off-by: Stephan Mueller <smueller@chronox.de>