LRNG: enhance performance of aux pool
The auxiliary pool is now a hash state like the per-CPU entropy pools.
This implies that adding data only requires a hash update operation. For
smaller data, this commonly only requires a copy of the data into the
hash look-aside storage. Therefore, the replacement of a full hash with
the hash update operation will deliver a significant performance
increase.

To support this hash state, an initialization of the hash is required
during boot time. Also, when data is to be obtained from the aux pool, a
hash final operation followed immediately by a hash init is required.

To support a switch of the hash, a function is added which closes the
hash state with the old hash and initialize the state with the new hash.
The message digest from the old hash is inserted into the new hash to
guarantee a chaining of the state.

Signed-off-by: Stephan Mueller <smueller@chronox.de>