LRNG: add CONFIG_LRNG_RUNTIME_MAX_WO_RESEED_CONFIG
The configuration option is intended for testing of the theshold of the
maximum number of generate operations without reseeding the LRNG with
full entropy.

It adds the lrng_drng.max_wo_reseed kernel command line option.

This option is not expected to be set for production kernels. In this
case, the entire additional code is folded into a noop.

Signed-off-by: Stephan Mueller <smueller@chronox.de>