LRNG: SP800-90C - reseed with security strength
SP800-90C (draft Q1 2021) section 6.2.1 requires the DRBG to be
initialized with s + 128 bits of entropy. This is already implemented by
the LRNG. Yet, for subsequent reseeds, the seed is only required to have
security strength bits s according to section 6.2.3.

This patch reduces the required entropy rate from s + 128 to s for fully
seeded DRNGs. In case no DRNG state is in scope allowing to make the
determination whether a fully seeded state is reached, the following
indicators are used:

- To determine whether the LRNG is operational, it is sufficient that the
first DRNG instance is fully seeded (which is indicated by the flag that
the LRNG is fully seeded).

- To determine the reseed threshold, the LRNG requires that all DRNGs
are fully seeded. This is to ensure that the LRNG-forced reseed trigger
is issued when having sufficient entropy.

The heart of the change is the helper lrng_get_seed_entropy_osr. All
callers are adapted accordingly.

Note, the invocation to lrng_set_entropy_thres in the lrng_switch.c
code is moved below the call to unset the DRNG fully seeded status
because the newly set threshold depends on whether the DRNG is fully
seeded or not.

The patch adds a test analyzing the correctness of the implementation.

Signed-off-by: Stephan Mueller <smueller@chronox.de>