LRNG: Dynamically allocate Jitter RNG
As the Jitter RNG is now initialized in device_initcall, we can drop the
static initialization of the Jitter RNG in favor of the dynamic
allocation. Thus, the instance provided with jent_lrng_entropy_collector
is not needed any more and will not be created with the patch series any
more.

Signed-off-by: Stephan Mueller <smueller@chronox.de>