LRNG: enable IRQ entropy source during compile time
Add a compile time switch to allow the IRQ entropy source to be enabled
or disabled. The code change moves several code lines into the location
where ifdef cluttering is limited.

With the disabling of the IRQ entropy source, the performance of the
general interrupt handler is increased by not incuring the penalties of
the LRNG operation. Yet, in this case the LRNG does not contain an
internal entropy source. Yet, as long as other entropy sources are
available providing appropriate entropy rates, the LRNG operates
accordingly.

Note, the boot-time steps of 32/128/256 are not triggered when the
interrupt entropy source is not available so far. Thus, it is
recommended to have one or more other entropy sources which collectively
must provide 256 bits of entropy or user space must cause a reseeding to
achieve a fully seeded level.

Signed-off-by: Stephan Mueller <smueller@chronox.de>