LRNG: update write wakeup threshold when new hash is loaded
The wakeup threshold should be at most as large as the message digest of
the used hash. If it is larger, a sleeper is always woken up to provide
entropy even though the entropy provider can never fill up to the
threshold as at most (message digest) bits of entropy can end up in the
auxiliary pool.

Yet, when the message digest is updated, the threshold is also updated:

- the maximum value allowed to be set via proc is adjusted to the new
  message digest size

- the current value is set to the value at most the size of the new
  message digest size

Signed-off-by: Stephan Mueller <smueller@chronox.de>