LRNG: add compression support for CPU entropy source
It is possible that CPU entropy sources do not provide full entropy. If
such entropy source is detected, the LRNG calculates the "multiplier"
for how much data is to be pulled from the entropy source to reach the
requested amount of entropy. This data is compressed with the hash
algorithm in use by the LRNG. The message digest is considered to have
full entropy.

The following multipliers are defined:

- PowerPC DARN: The PowerISA specification defines that the DARN
  instruction delivers 0.5 bits of entropy per data bit. This implies a
  multiplier of 2 to obtain twice the amount of data to satify the
  requested amount of entropy.

- Intel RDRAND: RDRAND is defined to be reseeded from RDSEED after
  satisfying 511 generate requests. Thus, the multiplier is 512 to
  ensure that the RDRAND data has an equivalent strenght as RDSEED.

Note, as soon as RISC-V CPU entropy source is made available in the
kernel, another multiplier is to be added. The RISC-V specification
defines a minimum entropy rate of 0.5 bits of entropy per data bit.

Signed-off-by: Stephan Mueller <smueller@chronox.de>