LRNG: archrandom - add SP800-90C ES oversampling
SP800-90C mandates that at least the last conditioning operation pull 64
more bits of entropy from the ES than is output from the conditioner. As
the CPU ES now has a conditioner that is applied when the ES delivers
not full entropy, the oversampling has to be applied as well. In this
case, the CPU ES is pullet 64 * data_multiplier more bits of data.

Signed-off-by: Stephan Mueller <smueller@chronox.de>