LRNG: Limit forced reseed to init DRNG If init DRNG reseed is due
If the inital DRNG reseeding threshold has been reached, limit a forced
reseed to the initial DRNG only. If this limit is not set, user space
could mount a DoS against the LRNG by forcing a reseed and manaing the
reseed goes to a NUMA-node DRNG that is not the initial DRNG. If the
inital DRNG would hit the hard ceiling at LRNG_DRNG_MAX_WITHOUT_RESEED
the entire LRNG will be offline in the worst case of this attack.

Signed-off-by: Stephan Mueller <smueller@chronox.de>