lrng: Achieve full ABI compliance
The LRNG has a number of entropy pools. Currently they all are reported
for /proc/sys/kernel/random/entropy_avail and .../poolsize as well as
RNDGETENTCNT.

However, user space callers may try to update the entropy up to the
point of reaching the maximum entropy amount. This is impossible with
the reporting of all entropy pools because external callers can never
supply entropy to make entropy_avail reaching poolsize. This is due to
the fact that calling users can only update one entropy but never the
others.

The change now makes the mentioned interfaces to only report the
auxiliary entropy pool (i.e. the pool the user can reach). This supports
the rngd implementation from rng-tools.

Signed-off-by: Stephan Mueller <smueller@chronox.de>