lrng: introduce DRNG with prediction resistance
The previous patch adding NTG.1 / FIPS IG 7.19 / D.K compliance caused a
reseeding before and after the use of a "regular" DRNG. This is a drag
on (a) performance of "regular" uses of the DRNG as well as on the
entropy sources as the reseedong is mandated to happen before and after
the use of the DRNG.

The current patch introduces a new DRNG instance that is solely used for
the aforementioned use case. This helps performance as "regular" DRNG
users are not affected (i.e. the regular DRNGs can be used in parallel
to the prediction resistance DRNG). In addition, the DRNG which always
used with prediction resistance only needs to be seeded before it is
used reducing the pressure on the entropy sources. Finally, using a
separate DRNG instance for the prediction resistance use case provides a
clear separation of use cases.

Signed-off-by: Stephan Mueller <smueller@chronox.de>