lrng: simplify forced seeding implementation
Revert the changes in lrng_es_mgr.* regarding the forced seeding
implementation. The forced seeding is only relating to the DRNG
management and is unrelated to the entropy sources. Thus, it must not be
present in lrng_es_mgr.*.

The reimplementation uses a trigger called by the interface functions
which triggers the forced seeding process.

The re-implementation now contains also a test case covering the
different aspects of the forced reseeding to verify it works correctly.
Also, the code allows a test interface to be compiled that can disable
the forced reseeding. This is used to update a different test in a
separate patch.

Signed-off-by: Stephan Mueller <smueller@chronox.de>