lrng: permanent health error and FIPS 140-3 compliance
To comply with the requirement of FIPS 140-3 that a health failure is
treated as a FIPS module error (A module error implies that all services
of the FIPS module must cease to operate. For the kernel crypto API that
is achieved by stopping the kernel, a situation that should be avoided
unless a catastrophic event happened). in case the LRNG is considered
part of a FIPS module, the following approach is taken:

1. The APT/RCT alpha values that are specific to the entropy source are
considered intermittent errors as defined by [#SP800-90B:2018] section
4.3 bullet 2. This means that in case a health test error is identified
for the entropy source, a reset of the entropy source is performed.
This implies that all potentially collected entropy is considered to be
removed, the health test is reset and a complete new power-on health
test is performed. The entropy source does not produce data in this
case and reports zero entropy to the LRNG. Yet, after a successful
power-on health test, the entropy source becomes available again to
serve entropy to the LRNG.

2. The LRNG implements a permanent APT/RCT failure with an alpha rate
of \alpha=2^{-60} that causes the kernel to stop. This permanent
failure is implemented by identifying two back-to-back intermittent
health failures. The chosen alpha value is considered appropriate
considering it is to be interpreted as a false-positive rate according
to [#SP800-90B:2018] section 4.2. Assuming the LRNG is used in larger
numbers of Linux kernels which run all concurrently, the false-positive
rate must be applied to all those systems running concurrently which
implies that the likelihood of a false-positive to occur is much higher.
For example, if 2^{20} Linux kernels with the LRNG execute concurrently,
the a false-positive permanent failure now happens with a probability
of 2^{20}\cdot\alpha=2^{-40} which again falls into the range of alpha
stipulated by [#SP800-90B:2018] section 4.3.

The intermittent failure will always trigger a reset of the offending
entropy source. The permanent failure is therefore only intended in
case the LRNG operates as part of a FIPS 140-3 module. If the LRNG is
outside of a FIPS module, the permanent failure does not need to be
considered. The reason is that with constant health test failures, the
entropy source is effectively disabled anyway as it does not produce
data. Yet, it has the capability to recover in case the environment
changes such that no health test errors occur.

When the LRNG operates as part of a FIPS module, the permanent health
failure, the entire FIPS module must cease to operate. One way to
ensure that is by setting the kernel boot time variable of
lrng_es_mgr.lrng_panic_on_permanent_health_failure to true. This will
panic the kernel when a permanent health failure is detected which
ensures that the FIPS module will cease to be operational.

Signed-off-by: Stephan Mueller <smueller@chronox.de>