internal ES: use two alpha values for intermittend/permanent health failure
According to SP800-90B, two health failures are allowed: the intermittend
and the permanent failure. So far, only the intermittent failure was
implemented. The permanent failure was achieved by resetting the entire
entropy source including its health test state and waiting for two or
more back-to-back health errors.

This approach is appropriate for RCT, but not for APT as APT has a
non-linear cutoff value. Thus, this patch implements 2 cutoff values
for both RCT/APT. This implies that the health state is left untouched
when an intermittent failure occurs. The noise sources are reset
and a new APT powerup-self test is performed. Yet, whith the unchanged
health test state, the counting of failures continues until a permanent
failure is reached.

Any non-failing raw entropy value causes the health tests to reset.

A permanent failure causes the entire entropy source to reset including
its health state plus an error note is sent. In case
lrng_es_mgr.lrng_panic_on_permanent_health_failure is set to 1, the
kernel panics - which is needed if the LRNG is part of a FIPS 140
module.

Signed-off-by: Stephan Mueller <smueller@chronox.de>