jent ES: use common kernel crypto API interface
Bypassing of the kernel crypto API interface in the past was done do
initialize the Jitter RNG very very early in the boot cycle at the time
when the kernel crypto API was not yet available. As this approach has
changed and the Jitter RNG ES is initialized after the kernel crypto
API is available, use its interfaces.

This simplifies the code and eliminates the requirement of another
patch.

Signed-off-by: Stephan Mueller <smueller@chronox.de>